webapi/webapi-rawsockets-w3c-tests/rawsockets/TCPServerSocket_attribute.html

<!DOCTYPE html>
<!--
Copyright (c) 2013 Intel Corporation.

Redistribution and use in source and binary forms, with or without modification,
are permitted provided that the following conditions are met:

* Redistributions of works must retain the original copyright notice, this list
  of conditions and the following disclaimer.
* Redistributions in binary form must reproduce the original copyright notice,
  this list of conditions and the following disclaimer in the documentation
  and/or other materials provided with the distribution.
* Neither the name of Intel Corporation nor the names of its contributors
  may be used to endorse or promote products derived from this work without
  specific prior written permission.

THIS SOFTWARE IS PROVIDED BY INTEL CORPORATION "AS IS"
AND ANY EXPRESS OR IMPLIED WARRANTIES, INCLUDING, BUT NOT LIMITED TO, THE
IMPLIED WARRANTIES OF MERCHANTABILITY AND FITNESS FOR A PARTICULAR PURPOSE
ARE DISCLAIMED. IN NO EVENT SHALL INTEL CORPORATION BE LIABLE FOR ANY DIRECT,
INDIRECT, INCIDENTAL, SPECIAL, EXEMPLARY, OR CONSEQUENTIAL DAMAGES (INCLUDING,
BUT NOT LIMITED TO, PROCUREMENT OF SUBSTITUTE GOODS OR SERVICES; LOSS OF USE,
DATA, OR PROFITS; OR BUSINESS INTERRUPTION) HOWEVER CAUSED AND ON ANY THEORY
OF LIABILITY, WHETHER IN CONTRACT, STRICT LIABILITY, OR TORT (INCLUDING
NEGLIGENCE OR OTHERWISE) ARISING IN ANY WAY OUT OF THE USE OF THIS SOFTWARE,
EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.

Authors:
        Shentu.Jiazhen <jiazhenx.shentu@intel.com>

-->

<meta charset="utf-8">
<title>Raw Socket Test: TCPServerSocket attribute</title>
<link rel="author" title="Intel" href="http://www.intel.com">
<link rel="help" href="http://www.w3.org/2012/sysapps/raw-sockets/#attributes-2">
<script src="../resources/testharness.js"></script>
<script src="../resources/testharnessreport.js"></script>
<script src="./support/support.js"></script>
<div id="log"></div>
<script>

test(function () {
  assert_true(!!namespace, "TCPServerSocket interface support");
  var socket = new namespace.TCPServerSocket(socketOptions);
  assert_true(socket.addressReuse, "TCPServerSocket.addressReuse default is");
}, "Check if the default value of TCPServerSocket.addressReuse is true");

test(function () {
  assert_true(!!namespace, "TCPServerSocket interface support");
  var socket = new namespace.TCPServerSocket({"localAddress" :"127.0.0.1", "localPort" : 1234, "addressReuse": false});
  assert_false(socket.addressReuse, "TCPServerSocket.addressReuse");
}, "Check if the value of TCPServerSocket.addressReuse can be set by the options argument in the constructor");

test(function () {
  assert_true(!!namespace, "TCPServerSocket interface support");
  var socket = new namespace.TCPServerSocket(socketOptions);
  assert_equals(socket.localAddress, socketOptions.localAddress, "TCPServerSocket.localAddress");
}, "Check if TCPServerSocket.localAddress can be set by the constructor's options argument's valid localAddress member");

test(function () {
  assert_true(!!namespace, "TCPServerSocket interface support");
  var socket = new namespace.TCPServerSocket(socketOptions);
  assert_equals(socket.localPort, socketOptions.localPort, "TCPServerSocket.localPort");
}, "Check if TCPServerSocket.localPort can be set by the constructor's options argument's valid localPort member");

</script>
